Create a separate file for Like class
